src/ingestion/requirements.txt
functions-framework
Flask
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-pubsub
google-cloud-logging
google-cloud-firestore
requests
tenacity
sec-api
python-dateutil
pandas
pandas-ta
pandas-gbq
pyarrow
db-dtypes
beautifulsoup4